Added files on agent run like in NIKA
